+++ src/tabs/status.rs
@@ -7,7 +7,7 @@ use crate::{
         FileTreeItemKind,
     },
     keys::SharedKeyConfig,
-    queue::{Action, InternalEvent, Queue, ResetItem},
+    queue::{Action, InternalEvent, NeedsUpdate, Queue, ResetItem},
     strings, try_or_popup,
     ui::style::SharedTheme,
 };
@@ -468,6 +468,14 @@ impl Status {
         }
     }
 
+    fn undo_last_commit(&self) {
+        try_or_popup!(
+            self,
+            "undo commit failed:",
+            sync::utils::undo_last_commit(CWD)
+        );
+    }
+
     fn branch_compare(&mut self) {
         self.git_branch_state =
             self.git_branch_name.last().and_then(|branch| {
@@ -580,6 +588,12 @@ impl Component for Status {
             ));
 
             out.push(CommandInfo::new(
+                strings::commands::undo_commit(&self.key_config),
+                true,
+                !focus_on_diff,
+            ));
+
+            out.push(CommandInfo::new(
                 strings::commands::abort_merge(&self.key_config),
                 true,
                 Self::can_abort_merge() || force_all,
@@ -686,6 +700,14 @@ impl Component for Status {
                     && !self.is_focus_on_diff()
                 {
                     self.pull();
+                    Ok(EventState::Consumed)
+                } else if k == self.key_config.undo_commit
+                    && !self.is_focus_on_diff()
+                {
+                    self.undo_last_commit();
+                    self.queue.borrow_mut().push_back(
+                        InternalEvent::Update(NeedsUpdate::ALL),
+                    );
                     Ok(EventState::Consumed)
                 } else if k == self.key_config.abort_merge
                     && Self::can_abort_merge()